refactor: alterando nome do componente
