+++ src/App.vue
@@ -1,7 +1,7 @@
 <template>
   <n-message-provider>
     <RouterView />
-    <ButtonHome />
+    <ActionsButton />
   </n-message-provider>
 </template>
 
@@ -10,7 +10,7 @@ import { onMounted } from 'vue'
 import { RouterView } from 'vue-router'
 import { NMessageProvider } from 'naive-ui'
 import AuthService from '@/class/services/AuthService.ts'
-import ButtonHome from '@/components/Global/ButtonHome.vue'
+import ActionsButton from '@/components/Global/ActionsButton.vue'
 
 const authService = new AuthService()
